refactor: Replace colors component with task-edit class methods
