+++ src/components/task.js
@@ -1,4 +1,4 @@
-import {createElement} from './util.js';
+import {createElement} from './utils.js';
 
 class Task {
   constructor({text, dueDate, repeatingDays, tags, color, isFavorite, isArchive}) {
